Rename vasp.py to poscar.py & tests
